+++ citybikes/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
 __title__ = 'python-citybikes'
-__version__ = '0.1.4'
-__copyright__ = 'Copyright 2016 Lluís Esquerda'
+__version__ = '0.2.0'
+__copyright__ = 'Copyright 2021 Lluís Esquerda'
 
 __all__ = ['Client', 'Network', 'Station']